Improve interfaces
