+++ src/decorators/register-in-is.decorator.ts
@@ -2,8 +2,8 @@ import {
   registerInIsArgumentDecoratorType,
   registerInIsConstructorDecoratorType,
 } from '../types/decorators/register-in-is.decorator.type';
-import { RegisterInIsDecoratorInterface } from '../interfaces/decorators/register-in-is.decorator.interface';
-import { methods } from '../methods';
+import {RegisterInIsDecoratorInterface} from '../interfaces/decorators/register-in-is.decorator.interface';
+import {predefinedMethods} from '../methods';
 
 export function registerInIsDecorator(configuration?: RegisterInIsDecoratorInterface): any {
   return (constructor: registerInIsConstructorDecoratorType): void => {
@@ -25,7 +25,7 @@ export function registerInIsDecorator(configuration?: RegisterInIsDecoratorInter
       const object: registerInIsArgumentDecoratorType = new constructor();
       className = object.constructor.name;
     }
-    Object.defineProperty(methods, className, {
+    Object.defineProperty(predefinedMethods, className, {
       configurable: true,
       enumerable: true,
       writable: true,
