+++ src/interfaces/methods/HTML-elements.methods.interface.ts
@@ -1,0 +1,82 @@
+export interface HTMLElementsMethodsInterface {
+    HTMLDirectoryElement: HTMLDirectoryElement;
+    HTMLDocument: HTMLDocument;
+    HTMLFontElement: HTMLFontElement;
+    HTMLFrameElement: HTMLFrameElement;
+    HTMLMarqueeElement: HTMLMarqueeElement;
+    HTMLParamElement: HTMLParamElement;
+    HTMLTableDataCellElement: HTMLTableDataCellElement;
+    HTMLTableHeaderCellElement: HTMLTableHeaderCellElement;
+    HTMLAllCollection: HTMLAllCollection;
+    HTMLCollectionBase: HTMLCollectionBase;
+    HTMLHyperlinkElementUtils: HTMLHyperlinkElementUtils;
+    HTMLOrSVGElement: HTMLOrSVGElement;
+    HTMLAnchorElement: HTMLAnchorElement;
+    HTMLAreaElement: HTMLAreaElement;
+    HTMLAudioElement: HTMLAudioElement;
+    HTMLBRElement: HTMLBRElement;
+    HTMLBaseElement: HTMLBaseElement;
+    HTMLBodyElement: HTMLBodyElement;
+    HTMLButtonElement: HTMLButtonElement;
+    HTMLCanvasElement: HTMLCanvasElement;
+    HTMLCollection: HTMLCollection;
+    HTMLDListElement: HTMLDListElement;
+    HTMLDataElement: HTMLDataElement;
+    HTMLDataListElement: HTMLDataListElement;
+    HTMLDetailsElement: HTMLDetailsElement;
+    HTMLDialogElement: HTMLDialogElement;
+    HTMLDivElement: HTMLDivElement;
+    HTMLElement: HTMLElement;
+    HTMLEmbedElement: HTMLEmbedElement;
+    HTMLFieldSetElement: HTMLFieldSetElement;
+    HTMLFormControlsCollection: HTMLFormControlsCollection;
+    HTMLFormElement: HTMLFormElement;
+    HTMLHRElement: HTMLHRElement;
+    HTMLHeadElement: HTMLHeadElement;
+    HTMLHeadingElement: HTMLHeadingElement;
+    HTMLHtmlElement: HTMLHtmlElement;
+    HTMLIFrameElement: HTMLIFrameElement;
+    HTMLImageElement: HTMLImageElement;
+    HTMLInputElement: HTMLInputElement;
+    HTMLLIElement: HTMLLIElement;
+    HTMLLabelElement: HTMLLabelElement;
+    HTMLLegendElement: HTMLLegendElement;
+    HTMLLinkElement: HTMLLinkElement;
+    HTMLMapElement: HTMLMapElement;
+    HTMLMediaElement: HTMLMediaElement;
+    HTMLMenuElement: HTMLMenuElement;
+    HTMLMetaElement: HTMLMetaElement;
+    HTMLMeterElement: HTMLMeterElement;
+    HTMLModElement: HTMLModElement;
+    HTMLOListElement: HTMLOListElement;
+    HTMLObjectElement: HTMLObjectElement;
+    HTMLOptGroupElement: HTMLOptGroupElement;
+    HTMLOptionElement: HTMLOptionElement;
+    HTMLOptionsCollection: HTMLOptionsCollection;
+    HTMLOutputElement: HTMLOutputElement;
+    HTMLParagraphElement: HTMLParagraphElement;
+    HTMLPictureElement: HTMLPictureElement;
+    HTMLPreElement: HTMLPreElement;
+    HTMLProgressElement: HTMLProgressElement;
+    HTMLQuoteElement: HTMLQuoteElement;
+    HTMLScriptElement: HTMLScriptElement;
+    HTMLSelectElement: HTMLSelectElement;
+    HTMLSlotElement: HTMLSlotElement;
+    HTMLSourceElement: HTMLSourceElement;
+    HTMLSpanElement: HTMLSpanElement;
+    HTMLStyleElement: HTMLStyleElement;
+    HTMLTableCaptionElement: HTMLTableCaptionElement;
+    HTMLTableCellElement: HTMLTableCellElement;
+    HTMLTableColElement: HTMLTableColElement;
+    HTMLTableElement: HTMLTableElement;
+    HTMLTableRowElement: HTMLTableRowElement;
+    HTMLTableSectionElement: HTMLTableSectionElement;
+    HTMLTemplateElement: HTMLTemplateElement;
+    HTMLTextAreaElement: HTMLTextAreaElement;
+    HTMLTimeElement: HTMLTimeElement;
+    HTMLTitleElement: HTMLTitleElement;
+    HTMLTrackElement: HTMLTrackElement;
+    HTMLUListElement: HTMLUListElement;
+    HTMLUnknownElement: HTMLUnknownElement;
+    HTMLVideoElement: HTMLVideoElement;
+}
