+++ src/interfaces/methods/external.methods.interface.ts
@@ -1,0 +1,844 @@
+export interface ExternalMethodsInterface {
+    Map: typeof Map;
+    String: string;
+    Date: Date;
+    Set: typeof Set;
+    URIError: URIError;
+    RegExp: RegExp;
+    WeakSet: typeof WeakSet;
+    WeakMap: typeof WeakMap;
+    DataView: DataView;
+    Float32Array: Float32Array;
+    Int32Array: Int32Array;
+    Uint8ClampedArray: Uint8ClampedArray;
+    Int8Array: Int8Array;
+    Uint8Array: Uint8Array;
+    Int16Array: Int16Array;
+    Uint16Array: Uint16Array;
+    Uint32Array: Uint32Array;
+    Float64Array: Float64Array;
+    BigInt64Array: BigInt64Array;
+    BigUint64Array: BigUint64Array;
+    RangeError: RangeError;
+    DOMException: DOMException;
+    Error: Error;
+    EvalError: EvalError;
+    ReferenceError: ReferenceError;
+    SyntaxError: SyntaxError;
+    TypeError: TypeError;
+    Algorithm: Algorithm;
+    AssignedNodesOptions: AssignedNodesOptions;
+    AudioBufferOptions: AudioBufferOptions;
+    AudioBufferSourceOptions: AudioBufferSourceOptions;
+    AudioConfiguration: AudioConfiguration;
+    AudioContextOptions: AudioContextOptions;
+    AudioNodeOptions: AudioNodeOptions;
+    AudioTimestamp: AudioTimestamp;
+    AuthenticationExtensionsClientInputs: AuthenticationExtensionsClientInputs;
+    AuthenticationExtensionsClientOutputs: AuthenticationExtensionsClientOutputs;
+    AuthenticatorSelectionCriteria: AuthenticatorSelectionCriteria;
+    BlobEventInit: BlobEventInit;
+    BlobPropertyBag: BlobPropertyBag;
+    CSSStyleSheetInit: CSSStyleSheetInit;
+    CacheQueryOptions: CacheQueryOptions;
+    CanvasRenderingContext2DSettings: CanvasRenderingContext2DSettings;
+    ClientQueryOptions: ClientQueryOptions;
+    ClipboardItemOptions: ClipboardItemOptions;
+    ComputedKeyframe: ComputedKeyframe;
+    ConstantSourceOptions: ConstantSourceOptions;
+    ConstrainBooleanParameters: ConstrainBooleanParameters;
+    ConstrainDOMStringParameters: ConstrainDOMStringParameters;
+    CredentialCreationOptions: CredentialCreationOptions;
+    CredentialPropertiesOutput: CredentialPropertiesOutput;
+    CredentialRequestOptions: CredentialRequestOptions;
+    CryptoKeyPair: CryptoKeyPair;
+    DOMMatrix2DInit: DOMMatrix2DInit;
+    DOMPointInit: DOMPointInit;
+    DOMQuadInit: DOMQuadInit;
+    DOMRectInit: DOMRectInit;
+    DeviceMotionEventAccelerationInit: DeviceMotionEventAccelerationInit;
+    DeviceMotionEventRotationRateInit: DeviceMotionEventRotationRateInit;
+    DisplayMediaStreamOptions: DisplayMediaStreamOptions;
+    DocumentTimelineOptions: DocumentTimelineOptions;
+    DoubleRange: DoubleRange;
+    EffectTiming: EffectTiming;
+    ElementCreationOptions: ElementCreationOptions;
+    ElementDefinitionOptions: ElementDefinitionOptions;
+    EventInit: EventInit;
+    EventListenerOptions: EventListenerOptions;
+    EventSourceInit: EventSourceInit;
+    FileSystemFlags: FileSystemFlags;
+    FileSystemGetDirectoryOptions: FileSystemGetDirectoryOptions;
+    FileSystemGetFileOptions: FileSystemGetFileOptions;
+    FileSystemRemoveOptions: FileSystemRemoveOptions;
+    FocusOptions: FocusOptions;
+    FontFaceDescriptors: FontFaceDescriptors;
+    FullscreenOptions: FullscreenOptions;
+    GetAnimationsOptions: GetAnimationsOptions;
+    GetNotificationOptions: GetNotificationOptions;
+    GetRootNodeOptions: GetRootNodeOptions;
+    IDBDatabaseInfo: IDBDatabaseInfo;
+    IDBIndexParameters: IDBIndexParameters;
+    IDBObjectStoreParameters: IDBObjectStoreParameters;
+    IDBTransactionOptions: IDBTransactionOptions;
+    IdleRequestOptions: IdleRequestOptions;
+    ImageBitmapOptions: ImageBitmapOptions;
+    ImageBitmapRenderingContextSettings: ImageBitmapRenderingContextSettings;
+    ImageDataSettings: ImageDataSettings;
+    ImportMeta: ImportMeta;
+    IntersectionObserverEntryInit: IntersectionObserverEntryInit;
+    IntersectionObserverInit: IntersectionObserverInit;
+    JsonWebKey: JsonWebKey;
+    KeyAlgorithm: KeyAlgorithm;
+    Keyframe: Keyframe;
+    LockInfo: LockInfo;
+    LockManagerSnapshot: LockManagerSnapshot;
+    LockOptions: LockOptions;
+    MediaCapabilitiesInfo: MediaCapabilitiesInfo;
+    MediaConfiguration: MediaConfiguration;
+    MediaElementAudioSourceOptions: MediaElementAudioSourceOptions;
+    MediaImage: MediaImage;
+    MediaKeySystemConfiguration: MediaKeySystemConfiguration;
+    MediaKeySystemMediaCapability: MediaKeySystemMediaCapability;
+    MediaMetadataInit: MediaMetadataInit;
+    MediaPositionState: MediaPositionState;
+    MediaRecorderOptions: MediaRecorderOptions;
+    MediaSessionActionDetails: MediaSessionActionDetails;
+    MediaStreamAudioSourceOptions: MediaStreamAudioSourceOptions;
+    MediaStreamConstraints: MediaStreamConstraints;
+    MediaTrackCapabilities: MediaTrackCapabilities;
+    MediaTrackConstraintSet: MediaTrackConstraintSet;
+    MediaTrackSettings: MediaTrackSettings;
+    MediaTrackSupportedConstraints: MediaTrackSupportedConstraints;
+    MutationObserverInit: MutationObserverInit;
+    NavigationPreloadState: NavigationPreloadState;
+    NotificationAction: NotificationAction;
+    NotificationOptions: NotificationOptions;
+    OfflineAudioContextOptions: OfflineAudioContextOptions;
+    OptionalEffectTiming: OptionalEffectTiming;
+    PaymentCurrencyAmount: PaymentCurrencyAmount;
+    PaymentDetailsBase: PaymentDetailsBase;
+    PaymentDetailsModifier: PaymentDetailsModifier;
+    PaymentItem: PaymentItem;
+    PaymentMethodData: PaymentMethodData;
+    PaymentValidationErrors: PaymentValidationErrors;
+    PerformanceMarkOptions: PerformanceMarkOptions;
+    PerformanceMeasureOptions: PerformanceMeasureOptions;
+    PerformanceObserverInit: PerformanceObserverInit;
+    PeriodicWaveConstraints: PeriodicWaveConstraints;
+    PermissionDescriptor: PermissionDescriptor;
+    PositionOptions: PositionOptions;
+    PropertyIndexedKeyframes: PropertyIndexedKeyframes;
+    PublicKeyCredentialCreationOptions: PublicKeyCredentialCreationOptions;
+    PublicKeyCredentialDescriptor: PublicKeyCredentialDescriptor;
+    PublicKeyCredentialEntity: PublicKeyCredentialEntity;
+    PublicKeyCredentialParameters: PublicKeyCredentialParameters;
+    PublicKeyCredentialRequestOptions: PublicKeyCredentialRequestOptions;
+    PushSubscriptionJSON: PushSubscriptionJSON;
+    PushSubscriptionOptionsInit: PushSubscriptionOptionsInit;
+    QueuingStrategyInit: QueuingStrategyInit;
+    RTCCertificateExpiration: RTCCertificateExpiration;
+    RTCConfiguration: RTCConfiguration;
+    RTCDataChannelInit: RTCDataChannelInit;
+    RTCDtlsFingerprint: RTCDtlsFingerprint;
+    RTCEncodedAudioFrameMetadata: RTCEncodedAudioFrameMetadata;
+    RTCEncodedVideoFrameMetadata: RTCEncodedVideoFrameMetadata;
+    RTCErrorInit: RTCErrorInit;
+    RTCIceCandidateInit: RTCIceCandidateInit;
+    RTCIceServer: RTCIceServer;
+    RTCLocalSessionDescriptionInit: RTCLocalSessionDescriptionInit;
+    RTCOfferAnswerOptions: RTCOfferAnswerOptions;
+    RTCRtcpParameters: RTCRtcpParameters;
+    RTCRtpCapabilities: RTCRtpCapabilities;
+    RTCRtpCodecCapability: RTCRtpCodecCapability;
+    RTCRtpCodecParameters: RTCRtpCodecParameters;
+    RTCRtpCodingParameters: RTCRtpCodingParameters;
+    RTCRtpContributingSource: RTCRtpContributingSource;
+    RTCRtpHeaderExtensionCapability: RTCRtpHeaderExtensionCapability;
+    RTCRtpHeaderExtensionParameters: RTCRtpHeaderExtensionParameters;
+    RTCRtpParameters: RTCRtpParameters;
+    RTCRtpTransceiverInit: RTCRtpTransceiverInit;
+    RTCSessionDescriptionInit: RTCSessionDescriptionInit;
+    RTCStats: RTCStats;
+    ReadableStreamGetReaderOptions: ReadableStreamGetReaderOptions;
+    RegistrationOptions: RegistrationOptions;
+    RequestInit: RequestInit;
+    ResizeObserverOptions: ResizeObserverOptions;
+    ResponseInit: ResponseInit;
+    RsaOtherPrimesInfo: RsaOtherPrimesInfo;
+    SVGBoundingBoxOptions: SVGBoundingBoxOptions;
+    ScrollOptions: ScrollOptions;
+    ShadowRootInit: ShadowRootInit;
+    ShareData: ShareData;
+    StaticRangeInit: StaticRangeInit;
+    StorageEstimate: StorageEstimate;
+    StreamPipeOptions: StreamPipeOptions;
+    StructuredSerializeOptions: StructuredSerializeOptions;
+    TextDecodeOptions: TextDecodeOptions;
+    TextDecoderOptions: TextDecoderOptions;
+    TextEncoderEncodeIntoResult: TextEncoderEncodeIntoResult;
+    TouchInit: TouchInit;
+    ULongRange: ULongRange;
+    UnderlyingByteSource: UnderlyingByteSource;
+    ValidityStateFlags: ValidityStateFlags;
+    VideoColorSpaceInit: VideoColorSpaceInit;
+    VideoConfiguration: VideoConfiguration;
+    VideoFrameCallbackMetadata: VideoFrameCallbackMetadata;
+    WebGLContextAttributes: WebGLContextAttributes;
+    WorkerOptions: WorkerOptions;
+    WorkletOptions: WorkletOptions;
+    ANGLE_instanced_arrays: ANGLE_instanced_arrays;
+    ARIAMixin: ARIAMixin;
+    AbortController: AbortController;
+    AbstractRange: AbstractRange;
+    AbstractWorker: AbstractWorker;
+    Animatable: Animatable;
+    AnimationEffect: AnimationEffect;
+    AnimationFrameProvider: AnimationFrameProvider;
+    AnimationTimeline: AnimationTimeline;
+    AudioBuffer: AudioBuffer;
+    AudioListener: AudioListener;
+    AudioParam: AudioParam;
+    AuthenticatorResponse: AuthenticatorResponse;
+    BarProp: BarProp;
+    Blob: Blob;
+    Body: Body;
+    CSSRule: CSSRule;
+    CSSRuleList: CSSRuleList;
+    CSSStyleDeclaration: CSSStyleDeclaration;
+    Cache: Cache;
+    CacheStorage: CacheStorage;
+    CanvasCompositing: CanvasCompositing;
+    CanvasDrawImage: CanvasDrawImage;
+    CanvasDrawPath: CanvasDrawPath;
+    CanvasFillStrokeStyles: CanvasFillStrokeStyles;
+    CanvasFilters: CanvasFilters;
+    CanvasGradient: CanvasGradient;
+    CanvasImageData: CanvasImageData;
+    CanvasImageSmoothing: CanvasImageSmoothing;
+    CanvasPath: CanvasPath;
+    CanvasPathDrawingStyles: CanvasPathDrawingStyles;
+    CanvasPattern: CanvasPattern;
+    CanvasRect: CanvasRect;
+    CanvasShadowStyles: CanvasShadowStyles;
+    CanvasState: CanvasState;
+    CanvasText: CanvasText;
+    CanvasTextDrawingStyles: CanvasTextDrawingStyles;
+    CanvasTransform: CanvasTransform;
+    CanvasUserInterface: CanvasUserInterface;
+    ClipboardItem: ClipboardItem;
+    Credential: Credential;
+    CredentialsContainer: CredentialsContainer;
+    Crypto: Crypto;
+    CryptoKey: CryptoKey;
+    CustomElementRegistry: CustomElementRegistry;
+    DOMImplementation: DOMImplementation;
+    DOMMatrixReadOnly: DOMMatrixReadOnly;
+    DOMParser: DOMParser;
+    DOMPointReadOnly: DOMPointReadOnly;
+    DOMQuad: DOMQuad;
+    DOMRectList: DOMRectList;
+    DOMRectReadOnly: DOMRectReadOnly;
+    DOMStringList: DOMStringList;
+    DOMTokenList: DOMTokenList;
+    DataTransfer: DataTransfer;
+    DataTransferItem: DataTransferItem;
+    DataTransferItemList: DataTransferItemList;
+    DeviceMotionEventAcceleration: DeviceMotionEventAcceleration;
+    DeviceMotionEventRotationRate: DeviceMotionEventRotationRate;
+    DocumentAndElementEventHandlers: DocumentAndElementEventHandlers;
+    DocumentOrShadowRoot: DocumentOrShadowRoot;
+    EXT_blend_minmax: EXT_blend_minmax;
+    EXT_color_buffer_float: EXT_color_buffer_float;
+    EXT_color_buffer_half_float: EXT_color_buffer_half_float;
+    EXT_float_blend: EXT_float_blend;
+    EXT_frag_depth: EXT_frag_depth;
+    EXT_sRGB: EXT_sRGB;
+    EXT_shader_texture_lod: EXT_shader_texture_lod;
+    EXT_texture_compression_bptc: EXT_texture_compression_bptc;
+    EXT_texture_compression_rgtc: EXT_texture_compression_rgtc;
+    EXT_texture_filter_anisotropic: EXT_texture_filter_anisotropic;
+    EXT_texture_norm16: EXT_texture_norm16;
+    ElementCSSInlineStyle: ElementCSSInlineStyle;
+    ElementContentEditable: ElementContentEditable;
+    Event: Event;
+    EventCounts: EventCounts;
+    EventListener: EventListener;
+    EventListenerObject: EventListenerObject;
+    EventTarget: EventTarget;
+    External: External;
+    FileList: FileList;
+    FileSystem: FileSystem;
+    FileSystemDirectoryReader: FileSystemDirectoryReader;
+    FileSystemEntry: FileSystemEntry;
+    FileSystemHandle: FileSystemHandle;
+    FontFace: FontFace;
+    FontFaceSource: FontFaceSource;
+    FormData: FormData;
+    Gamepad: Gamepad;
+    GamepadButton: GamepadButton;
+    GamepadHapticActuator: GamepadHapticActuator;
+    GenericTransformStream: GenericTransformStream;
+    Geolocation: Geolocation;
+    GeolocationCoordinates: GeolocationCoordinates;
+    GeolocationPosition: GeolocationPosition;
+    GeolocationPositionError: GeolocationPositionError;
+    GlobalEventHandlers: GlobalEventHandlers;
+    Headers: Headers;
+    History: History;
+    IDBCursor: IDBCursor;
+    IDBFactory: IDBFactory;
+    IDBIndex: IDBIndex;
+    IDBKeyRange: IDBKeyRange;
+    IDBObjectStore: IDBObjectStore;
+    IdleDeadline: IdleDeadline;
+    ImageBitmap: ImageBitmap;
+    ImageBitmapRenderingContext: ImageBitmapRenderingContext;
+    ImageData: ImageData;
+    InnerHTML: InnerHTML;
+    IntersectionObserver: IntersectionObserver;
+    IntersectionObserverEntry: IntersectionObserverEntry;
+    KHR_parallel_shader_compile: KHR_parallel_shader_compile;
+    LinkStyle: LinkStyle;
+    Location: Location;
+    Lock: Lock;
+    LockManager: LockManager;
+    MediaCapabilities: MediaCapabilities;
+    MediaDeviceInfo: MediaDeviceInfo;
+    MediaError: MediaError;
+    MediaKeySystemAccess: MediaKeySystemAccess;
+    MediaKeys: MediaKeys;
+    MediaList: MediaList;
+    MediaMetadata: MediaMetadata;
+    MediaSession: MediaSession;
+    MessageChannel: MessageChannel;
+    MimeType: MimeType;
+    MimeTypeArray: MimeTypeArray;
+    MutationObserver: MutationObserver;
+    MutationRecord: MutationRecord;
+    NavigationPreloadManager: NavigationPreloadManager;
+    NavigatorAutomationInformation: NavigatorAutomationInformation;
+    NavigatorConcurrentHardware: NavigatorConcurrentHardware;
+    NavigatorContentUtils: NavigatorContentUtils;
+    NavigatorCookies: NavigatorCookies;
+    NavigatorID: NavigatorID;
+    NavigatorLanguage: NavigatorLanguage;
+    NavigatorLocks: NavigatorLocks;
+    NavigatorOnLine: NavigatorOnLine;
+    NavigatorPlugins: NavigatorPlugins;
+    NavigatorStorage: NavigatorStorage;
+    NodeIterator: NodeIterator;
+    NodeList: NodeList;
+    NonDocumentTypeChildNode: NonDocumentTypeChildNode;
+    NonElementParentNode: NonElementParentNode;
+    OES_draw_buffers_indexed: OES_draw_buffers_indexed;
+    OES_element_index_uint: OES_element_index_uint;
+    OES_standard_derivatives: OES_standard_derivatives;
+    OES_texture_float: OES_texture_float;
+    OES_texture_float_linear: OES_texture_float_linear;
+    OES_texture_half_float: OES_texture_half_float;
+    OES_texture_half_float_linear: OES_texture_half_float_linear;
+    OES_vertex_array_object: OES_vertex_array_object;
+    OVR_multiview2: OVR_multiview2;
+    PerformanceEntry: PerformanceEntry;
+    PerformanceNavigation: PerformanceNavigation;
+    PerformanceObserver: PerformanceObserver;
+    PerformanceObserverEntryList: PerformanceObserverEntryList;
+    PerformanceServerTiming: PerformanceServerTiming;
+    PerformanceTiming: PerformanceTiming;
+    PeriodicWave: PeriodicWave;
+    Permissions: Permissions;
+    Plugin: Plugin;
+    PluginArray: PluginArray;
+    PushManager: PushManager;
+    PushSubscription: PushSubscription;
+    PushSubscriptionOptions: PushSubscriptionOptions;
+    RTCCertificate: RTCCertificate;
+    RTCEncodedAudioFrame: RTCEncodedAudioFrame;
+    RTCEncodedVideoFrame: RTCEncodedVideoFrame;
+    RTCIceCandidate: RTCIceCandidate;
+    RTCRtpReceiver: RTCRtpReceiver;
+    RTCRtpSender: RTCRtpSender;
+    RTCRtpTransceiver: RTCRtpTransceiver;
+    RTCSessionDescription: RTCSessionDescription;
+    RTCStatsReport: RTCStatsReport;
+    ReadableByteStreamController: ReadableByteStreamController;
+    ReadableStreamBYOBRequest: ReadableStreamBYOBRequest;
+    ReadableStreamGenericReader: ReadableStreamGenericReader;
+    ResizeObserver: ResizeObserver;
+    ResizeObserverEntry: ResizeObserverEntry;
+    ResizeObserverSize: ResizeObserverSize;
+    SVGAngle: SVGAngle;
+    SVGAnimatedAngle: SVGAnimatedAngle;
+    SVGAnimatedBoolean: SVGAnimatedBoolean;
+    SVGAnimatedEnumeration: SVGAnimatedEnumeration;
+    SVGAnimatedInteger: SVGAnimatedInteger;
+    SVGAnimatedLength: SVGAnimatedLength;
+    SVGAnimatedLengthList: SVGAnimatedLengthList;
+    SVGAnimatedNumber: SVGAnimatedNumber;
+    SVGAnimatedNumberList: SVGAnimatedNumberList;
+    SVGAnimatedPoints: SVGAnimatedPoints;
+    SVGAnimatedPreserveAspectRatio: SVGAnimatedPreserveAspectRatio;
+    SVGAnimatedRect: SVGAnimatedRect;
+    SVGAnimatedString: SVGAnimatedString;
+    SVGAnimatedTransformList: SVGAnimatedTransformList;
+    SVGFilterPrimitiveStandardAttributes: SVGFilterPrimitiveStandardAttributes;
+    SVGFitToViewBox: SVGFitToViewBox;
+    SVGLength: SVGLength;
+    SVGLengthList: SVGLengthList;
+    SVGNumber: SVGNumber;
+    SVGNumberList: SVGNumberList;
+    SVGPointList: SVGPointList;
+    SVGPreserveAspectRatio: SVGPreserveAspectRatio;
+    SVGStringList: SVGStringList;
+    SVGTests: SVGTests;
+    SVGTransform: SVGTransform;
+    SVGTransformList: SVGTransformList;
+    SVGURIReference: SVGURIReference;
+    SVGUnitTypes: SVGUnitTypes;
+    Screen: Screen;
+    Selection: Selection;
+    Slottable: Slottable;
+    SpeechRecognitionAlternative: SpeechRecognitionAlternative;
+    SpeechRecognitionResult: SpeechRecognitionResult;
+    SpeechRecognitionResultList: SpeechRecognitionResultList;
+    SpeechSynthesisVoice: SpeechSynthesisVoice;
+    Storage: Storage;
+    StorageManager: StorageManager;
+    StyleMedia: StyleMedia;
+    StyleSheet: StyleSheet;
+    StyleSheetList: StyleSheetList;
+    SubtleCrypto: SubtleCrypto;
+    TextDecoderCommon: TextDecoderCommon;
+    TextEncoderCommon: TextEncoderCommon;
+    TextMetrics: TextMetrics;
+    TextTrackCueList: TextTrackCueList;
+    TimeRanges: TimeRanges;
+    Touch: Touch;
+    TouchList: TouchList;
+    TreeWalker: TreeWalker;
+    URL: URL;
+    URLSearchParams: URLSearchParams;
+    VTTRegion: VTTRegion;
+    ValidityState: ValidityState;
+    VideoColorSpace: VideoColorSpace;
+    VideoPlaybackQuality: VideoPlaybackQuality;
+    WEBGL_color_buffer_float: WEBGL_color_buffer_float;
+    WEBGL_compressed_texture_astc: WEBGL_compressed_texture_astc;
+    WEBGL_compressed_texture_etc: WEBGL_compressed_texture_etc;
+    WEBGL_compressed_texture_etc1: WEBGL_compressed_texture_etc1;
+    WEBGL_compressed_texture_s3tc: WEBGL_compressed_texture_s3tc;
+    WEBGL_compressed_texture_s3tc_srgb: WEBGL_compressed_texture_s3tc_srgb;
+    WEBGL_debug_renderer_info: WEBGL_debug_renderer_info;
+    WEBGL_debug_shaders: WEBGL_debug_shaders;
+    WEBGL_depth_texture: WEBGL_depth_texture;
+    WEBGL_draw_buffers: WEBGL_draw_buffers;
+    WEBGL_lose_context: WEBGL_lose_context;
+    WEBGL_multi_draw: WEBGL_multi_draw;
+    WebGL2RenderingContextBase: WebGL2RenderingContextBase;
+    WebGL2RenderingContextOverloads: WebGL2RenderingContextOverloads;
+    WebGLActiveInfo: WebGLActiveInfo;
+    WebGLBuffer: WebGLBuffer;
+    WebGLFramebuffer: WebGLFramebuffer;
+    WebGLProgram: WebGLProgram;
+    WebGLQuery: WebGLQuery;
+    WebGLRenderbuffer: WebGLRenderbuffer;
+    WebGLRenderingContextBase: WebGLRenderingContextBase;
+    WebGLRenderingContextOverloads: WebGLRenderingContextOverloads;
+    WebGLSampler: WebGLSampler;
+    WebGLShader: WebGLShader;
+    WebGLShaderPrecisionFormat: WebGLShaderPrecisionFormat;
+    WebGLSync: WebGLSync;
+    WebGLTexture: WebGLTexture;
+    WebGLTransformFeedback: WebGLTransformFeedback;
+    WebGLUniformLocation: WebGLUniformLocation;
+    WebGLVertexArrayObject: WebGLVertexArrayObject;
+    WebGLVertexArrayObjectOES: WebGLVertexArrayObjectOES;
+    WindowEventHandlers: WindowEventHandlers;
+    WindowLocalStorage: WindowLocalStorage;
+    WindowOrWorkerGlobalScope: WindowOrWorkerGlobalScope;
+    WindowSessionStorage: WindowSessionStorage;
+    Worklet: Worklet;
+    WritableStreamDefaultController: WritableStreamDefaultController;
+    XMLSerializer: XMLSerializer;
+    XPathEvaluatorBase: XPathEvaluatorBase;
+    XPathExpression: XPathExpression;
+    XPathResult: XPathResult;
+    XSLTProcessor: XSLTProcessor;
+    BlobCallback: BlobCallback;
+    CustomElementConstructor: CustomElementConstructor;
+    DecodeErrorCallback: DecodeErrorCallback;
+    DecodeSuccessCallback: DecodeSuccessCallback;
+    ErrorCallback: ErrorCallback;
+    FileCallback: FileCallback;
+    FileSystemEntriesCallback: FileSystemEntriesCallback;
+    FileSystemEntryCallback: FileSystemEntryCallback;
+    FrameRequestCallback: FrameRequestCallback;
+    FunctionStringCallback: FunctionStringCallback;
+    IdleRequestCallback: IdleRequestCallback;
+    IntersectionObserverCallback: IntersectionObserverCallback;
+    LockGrantedCallback: LockGrantedCallback;
+    MediaSessionActionHandler: MediaSessionActionHandler;
+    MutationCallback: MutationCallback;
+    NotificationPermissionCallback: NotificationPermissionCallback;
+    OnBeforeUnloadEventHandlerNonNull: OnBeforeUnloadEventHandlerNonNull;
+    OnErrorEventHandlerNonNull: OnErrorEventHandlerNonNull;
+    PerformanceObserverCallback: PerformanceObserverCallback;
+    PositionCallback: PositionCallback;
+    PositionErrorCallback: PositionErrorCallback;
+    RTCPeerConnectionErrorCallback: RTCPeerConnectionErrorCallback;
+    RTCSessionDescriptionCallback: RTCSessionDescriptionCallback;
+    RemotePlaybackAvailabilityCallback: RemotePlaybackAvailabilityCallback;
+    ResizeObserverCallback: ResizeObserverCallback;
+    UnderlyingSinkAbortCallback: UnderlyingSinkAbortCallback;
+    UnderlyingSinkCloseCallback: UnderlyingSinkCloseCallback;
+    UnderlyingSinkStartCallback: UnderlyingSinkStartCallback;
+    UnderlyingSourceCancelCallback: UnderlyingSourceCancelCallback;
+    VideoFrameRequestCallback: VideoFrameRequestCallback;
+    VoidFunction: VoidFunction;
+    AddEventListenerOptions: AddEventListenerOptions;
+    AesCbcParams: AesCbcParams;
+    AesCtrParams: AesCtrParams;
+    AesDerivedKeyParams: AesDerivedKeyParams;
+    AesGcmParams: AesGcmParams;
+    AesKeyAlgorithm: AesKeyAlgorithm;
+    AesKeyGenParams: AesKeyGenParams;
+    AnalyserOptions: AnalyserOptions;
+    AnimationEventInit: AnimationEventInit;
+    AnimationPlaybackEventInit: AnimationPlaybackEventInit;
+    AudioProcessingEventInit: AudioProcessingEventInit;
+    AudioWorkletNodeOptions: AudioWorkletNodeOptions;
+    BiquadFilterOptions: BiquadFilterOptions;
+    ChannelMergerOptions: ChannelMergerOptions;
+    ChannelSplitterOptions: ChannelSplitterOptions;
+    ClipboardEventInit: ClipboardEventInit;
+    CloseEventInit: CloseEventInit;
+    CompositionEventInit: CompositionEventInit;
+    ComputedEffectTiming: ComputedEffectTiming;
+    ConstrainDoubleRange: ConstrainDoubleRange;
+    ConstrainULongRange: ConstrainULongRange;
+    ConvolverOptions: ConvolverOptions;
+    DOMMatrixInit: DOMMatrixInit;
+    DelayOptions: DelayOptions;
+    DeviceMotionEventInit: DeviceMotionEventInit;
+    DeviceOrientationEventInit: DeviceOrientationEventInit;
+    DragEventInit: DragEventInit;
+    DynamicsCompressorOptions: DynamicsCompressorOptions;
+    EcKeyAlgorithm: EcKeyAlgorithm;
+    EcKeyGenParams: EcKeyGenParams;
+    EcKeyImportParams: EcKeyImportParams;
+    EcdhKeyDeriveParams: EcdhKeyDeriveParams;
+    EcdsaParams: EcdsaParams;
+    ErrorEventInit: ErrorEventInit;
+    EventModifierInit: EventModifierInit;
+    FilePropertyBag: FilePropertyBag;
+    FocusEventInit: FocusEventInit;
+    FontFaceSetLoadEventInit: FontFaceSetLoadEventInit;
+    FormDataEventInit: FormDataEventInit;
+    GainOptions: GainOptions;
+    GamepadEventInit: GamepadEventInit;
+    HashChangeEventInit: HashChangeEventInit;
+    HkdfParams: HkdfParams;
+    HmacImportParams: HmacImportParams;
+    HmacKeyAlgorithm: HmacKeyAlgorithm;
+    HmacKeyGenParams: HmacKeyGenParams;
+    IDBVersionChangeEventInit: IDBVersionChangeEventInit;
+    IIRFilterOptions: IIRFilterOptions;
+    InputEventInit: InputEventInit;
+    KeyboardEventInit: KeyboardEventInit;
+    KeyframeAnimationOptions: KeyframeAnimationOptions;
+    KeyframeEffectOptions: KeyframeEffectOptions;
+    MediaCapabilitiesDecodingInfo: MediaCapabilitiesDecodingInfo;
+    MediaCapabilitiesEncodingInfo: MediaCapabilitiesEncodingInfo;
+    MediaDecodingConfiguration: MediaDecodingConfiguration;
+    MediaEncodingConfiguration: MediaEncodingConfiguration;
+    MediaEncryptedEventInit: MediaEncryptedEventInit;
+    MediaKeyMessageEventInit: MediaKeyMessageEventInit;
+    MediaQueryListEventInit: MediaQueryListEventInit;
+    MediaStreamTrackEventInit: MediaStreamTrackEventInit;
+    MediaTrackConstraints: MediaTrackConstraints;
+    MouseEventInit: MouseEventInit;
+    MultiCacheQueryOptions: MultiCacheQueryOptions;
+    OfflineAudioCompletionEventInit: OfflineAudioCompletionEventInit;
+    OscillatorOptions: OscillatorOptions;
+    PageTransitionEventInit: PageTransitionEventInit;
+    PannerOptions: PannerOptions;
+    PaymentDetailsInit: PaymentDetailsInit;
+    PaymentDetailsUpdate: PaymentDetailsUpdate;
+    PaymentMethodChangeEventInit: PaymentMethodChangeEventInit;
+    PaymentRequestUpdateEventInit: PaymentRequestUpdateEventInit;
+    Pbkdf2Params: Pbkdf2Params;
+    PeriodicWaveOptions: PeriodicWaveOptions;
+    PictureInPictureEventInit: PictureInPictureEventInit;
+    PointerEventInit: PointerEventInit;
+    PopStateEventInit: PopStateEventInit;
+    ProgressEventInit: ProgressEventInit;
+    PromiseRejectionEventInit: PromiseRejectionEventInit;
+    PublicKeyCredentialRpEntity: PublicKeyCredentialRpEntity;
+    PublicKeyCredentialUserEntity: PublicKeyCredentialUserEntity;
+    RTCAnswerOptions: RTCAnswerOptions;
+    RTCDTMFToneChangeEventInit: RTCDTMFToneChangeEventInit;
+    RTCDataChannelEventInit: RTCDataChannelEventInit;
+    RTCErrorEventInit: RTCErrorEventInit;
+    RTCIceCandidatePairStats: RTCIceCandidatePairStats;
+    RTCInboundRtpStreamStats: RTCInboundRtpStreamStats;
+    RTCOfferOptions: RTCOfferOptions;
+    RTCOutboundRtpStreamStats: RTCOutboundRtpStreamStats;
+    RTCPeerConnectionIceErrorEventInit: RTCPeerConnectionIceErrorEventInit;
+    RTCPeerConnectionIceEventInit: RTCPeerConnectionIceEventInit;
+    RTCReceivedRtpStreamStats: RTCReceivedRtpStreamStats;
+    RTCRtpEncodingParameters: RTCRtpEncodingParameters;
+    RTCRtpReceiveParameters: RTCRtpReceiveParameters;
+    RTCRtpSendParameters: RTCRtpSendParameters;
+    RTCRtpStreamStats: RTCRtpStreamStats;
+    RTCRtpSynchronizationSource: RTCRtpSynchronizationSource;
+    RTCSentRtpStreamStats: RTCSentRtpStreamStats;
+    RTCTrackEventInit: RTCTrackEventInit;
+    RTCTransportStats: RTCTransportStats;
+    RsaHashedImportParams: RsaHashedImportParams;
+    RsaHashedKeyAlgorithm: RsaHashedKeyAlgorithm;
+    RsaHashedKeyGenParams: RsaHashedKeyGenParams;
+    RsaKeyAlgorithm: RsaKeyAlgorithm;
+    RsaKeyGenParams: RsaKeyGenParams;
+    RsaOaepParams: RsaOaepParams;
+    RsaPssParams: RsaPssParams;
+    ScrollIntoViewOptions: ScrollIntoViewOptions;
+    ScrollToOptions: ScrollToOptions;
+    SecurityPolicyViolationEventInit: SecurityPolicyViolationEventInit;
+    SpeechSynthesisErrorEventInit: SpeechSynthesisErrorEventInit;
+    SpeechSynthesisEventInit: SpeechSynthesisEventInit;
+    StereoPannerOptions: StereoPannerOptions;
+    StorageEventInit: StorageEventInit;
+    SubmitEventInit: SubmitEventInit;
+    TouchEventInit: TouchEventInit;
+    TrackEventInit: TrackEventInit;
+    TransitionEventInit: TransitionEventInit;
+    UIEventInit: UIEventInit;
+    WaveShaperOptions: WaveShaperOptions;
+    WebGLContextEventInit: WebGLContextEventInit;
+    WheelEventInit: WheelEventInit;
+    WindowPostMessageOptions: WindowPostMessageOptions;
+    AbortSignal: AbortSignal;
+    AnalyserNode: AnalyserNode;
+    Animation: Animation;
+    AnimationEvent: AnimationEvent;
+    AnimationPlaybackEvent: AnimationPlaybackEvent;
+    Attr: Attr;
+    AudioBufferSourceNode: AudioBufferSourceNode;
+    AudioContext: AudioContext;
+    AudioDestinationNode: AudioDestinationNode;
+    AudioNode: AudioNode;
+    AudioProcessingEvent: AudioProcessingEvent;
+    AudioScheduledSourceNode: AudioScheduledSourceNode;
+    AudioWorklet: AudioWorklet;
+    AudioWorkletNode: AudioWorkletNode;
+    AuthenticatorAssertionResponse: AuthenticatorAssertionResponse;
+    AuthenticatorAttestationResponse: AuthenticatorAttestationResponse;
+    BaseAudioContext: BaseAudioContext;
+    BeforeUnloadEvent: BeforeUnloadEvent;
+    BiquadFilterNode: BiquadFilterNode;
+    BlobEvent: BlobEvent;
+    BroadcastChannel: BroadcastChannel;
+    CDATASection: CDATASection;
+    CSSAnimation: CSSAnimation;
+    CSSConditionRule: CSSConditionRule;
+    CSSContainerRule: CSSContainerRule;
+    CSSCounterStyleRule: CSSCounterStyleRule;
+    CSSFontFaceRule: CSSFontFaceRule;
+    CSSFontPaletteValuesRule: CSSFontPaletteValuesRule;
+    CSSGroupingRule: CSSGroupingRule;
+    CSSImportRule: CSSImportRule;
+    CSSKeyframeRule: CSSKeyframeRule;
+    CSSKeyframesRule: CSSKeyframesRule;
+    CSSLayerBlockRule: CSSLayerBlockRule;
+    CSSLayerStatementRule: CSSLayerStatementRule;
+    CSSMediaRule: CSSMediaRule;
+    CSSNamespaceRule: CSSNamespaceRule;
+    CSSPageRule: CSSPageRule;
+    CSSStyleRule: CSSStyleRule;
+    CSSStyleSheet: CSSStyleSheet;
+    CSSSupportsRule: CSSSupportsRule;
+    CSSTransition: CSSTransition;
+    CanvasCaptureMediaStreamTrack: CanvasCaptureMediaStreamTrack;
+    ChannelMergerNode: ChannelMergerNode;
+    ChannelSplitterNode: ChannelSplitterNode;
+    ChildNode: ChildNode;
+    ClientRect: ClientRect;
+    Clipboard: Clipboard;
+    ClipboardEvent: ClipboardEvent;
+    CloseEvent: CloseEvent;
+    Comment: Comment;
+    CompositionEvent: CompositionEvent;
+    ConstantSourceNode: ConstantSourceNode;
+    ConvolverNode: ConvolverNode;
+    CountQueuingStrategy: CountQueuingStrategy;
+    DOMMatrix: DOMMatrix;
+    DOMPoint: DOMPoint;
+    DOMRect: DOMRect;
+    DelayNode: DelayNode;
+    DeviceMotionEvent: DeviceMotionEvent;
+    DeviceOrientationEvent: DeviceOrientationEvent;
+    DocumentTimeline: DocumentTimeline;
+    DragEvent: DragEvent;
+    DynamicsCompressorNode: DynamicsCompressorNode;
+    ElementInternals: ElementInternals;
+    ErrorEvent: ErrorEvent;
+    EventSource: EventSource;
+    File: File;
+    FileReader: FileReader;
+    FileSystemDirectoryEntry: FileSystemDirectoryEntry;
+    FileSystemDirectoryHandle: FileSystemDirectoryHandle;
+    FileSystemFileEntry: FileSystemFileEntry;
+    FileSystemFileHandle: FileSystemFileHandle;
+    FocusEvent: FocusEvent;
+    FontFaceSet: FontFaceSet;
+    FontFaceSetLoadEvent: FontFaceSetLoadEvent;
+    FormDataEvent: FormDataEvent;
+    GainNode: GainNode;
+    GamepadEvent: GamepadEvent;
+    HashChangeEvent: HashChangeEvent;
+    IDBCursorWithValue: IDBCursorWithValue;
+    IDBDatabase: IDBDatabase;
+    IDBTransaction: IDBTransaction;
+    IDBVersionChangeEvent: IDBVersionChangeEvent;
+    IIRFilterNode: IIRFilterNode;
+    InputDeviceInfo: InputDeviceInfo;
+    InputEvent: InputEvent;
+    KeyboardEvent: KeyboardEvent;
+    KeyframeEffect: KeyframeEffect;
+    MediaDevices: MediaDevices;
+    MediaElementAudioSourceNode: MediaElementAudioSourceNode;
+    MediaEncryptedEvent: MediaEncryptedEvent;
+    MediaKeyMessageEvent: MediaKeyMessageEvent;
+    MediaKeySession: MediaKeySession;
+    MediaQueryList: MediaQueryList;
+    MediaQueryListEvent: MediaQueryListEvent;
+    MediaRecorder: MediaRecorder;
+    MediaSource: MediaSource;
+    MediaStream: MediaStream;
+    MediaStreamAudioDestinationNode: MediaStreamAudioDestinationNode;
+    MediaStreamAudioSourceNode: MediaStreamAudioSourceNode;
+    MediaStreamTrack: MediaStreamTrack;
+    MediaStreamTrackEvent: MediaStreamTrackEvent;
+    MessagePort: MessagePort;
+    MouseEvent: MouseEvent;
+    MutationEvent: MutationEvent;
+    Node: Node;
+    Notification: Notification;
+    OfflineAudioCompletionEvent: OfflineAudioCompletionEvent;
+    OfflineAudioContext: OfflineAudioContext;
+    OffscreenCanvas: OffscreenCanvas;
+    OscillatorNode: OscillatorNode;
+    OverconstrainedError: OverconstrainedError;
+    PageTransitionEvent: PageTransitionEvent;
+    PannerNode: PannerNode;
+    ParentNode: ParentNode;
+    Path2D: Path2D;
+    PaymentMethodChangeEvent: PaymentMethodChangeEvent;
+    PaymentRequest: PaymentRequest;
+    PaymentRequestUpdateEvent: PaymentRequestUpdateEvent;
+    PaymentResponse: PaymentResponse;
+    Performance: Performance;
+    PerformanceEventTiming: PerformanceEventTiming;
+    PerformanceMark: PerformanceMark;
+    PerformanceMeasure: PerformanceMeasure;
+    PerformanceNavigationTiming: PerformanceNavigationTiming;
+    PerformancePaintTiming: PerformancePaintTiming;
+    PerformanceResourceTiming: PerformanceResourceTiming;
+    PermissionStatus: PermissionStatus;
+    PictureInPictureEvent: PictureInPictureEvent;
+    PictureInPictureWindow: PictureInPictureWindow;
+    PointerEvent: PointerEvent;
+    PopStateEvent: PopStateEvent;
+    PromiseRejectionEvent: PromiseRejectionEvent;
+    PublicKeyCredential: PublicKeyCredential;
+    RTCDTMFSender: RTCDTMFSender;
+    RTCDTMFToneChangeEvent: RTCDTMFToneChangeEvent;
+    RTCDataChannel: RTCDataChannel;
+    RTCDataChannelEvent: RTCDataChannelEvent;
+    RTCDtlsTransport: RTCDtlsTransport;
+    RTCError: RTCError;
+    RTCErrorEvent: RTCErrorEvent;
+    RTCIceTransport: RTCIceTransport;
+    RTCPeerConnection: RTCPeerConnection;
+    RTCPeerConnectionIceErrorEvent: RTCPeerConnectionIceErrorEvent;
+    RTCPeerConnectionIceEvent: RTCPeerConnectionIceEvent;
+    RTCSctpTransport: RTCSctpTransport;
+    RTCTrackEvent: RTCTrackEvent;
+    RadioNodeList: RadioNodeList;
+    Range: Range;
+    ReadableStreamBYOBReader: ReadableStreamBYOBReader;
+    RemotePlayback: RemotePlayback;
+    Request: Request;
+    Response: Response;
+    SVGAnimateElement: SVGAnimateElement;
+    SVGAnimateMotionElement: SVGAnimateMotionElement;
+    SVGAnimateTransformElement: SVGAnimateTransformElement;
+    SVGCircleElement: SVGCircleElement;
+    SVGClipPathElement: SVGClipPathElement;
+    SVGComponentTransferFunctionElement: SVGComponentTransferFunctionElement;
+    SVGDefsElement: SVGDefsElement;
+    SVGDescElement: SVGDescElement;
+    SVGEllipseElement: SVGEllipseElement;
+    SVGFEDistantLightElement: SVGFEDistantLightElement;
+    SVGFEFuncAElement: SVGFEFuncAElement;
+    SVGFEFuncBElement: SVGFEFuncBElement;
+    SVGFEFuncGElement: SVGFEFuncGElement;
+    SVGFEFuncRElement: SVGFEFuncRElement;
+    SVGFEMergeNodeElement: SVGFEMergeNodeElement;
+    SVGFEPointLightElement: SVGFEPointLightElement;
+    SVGFESpotLightElement: SVGFESpotLightElement;
+    SVGForeignObjectElement: SVGForeignObjectElement;
+    SVGGElement: SVGGElement;
+    SVGGeometryElement: SVGGeometryElement;
+    SVGLineElement: SVGLineElement;
+    SVGLinearGradientElement: SVGLinearGradientElement;
+    SVGMaskElement: SVGMaskElement;
+    SVGMetadataElement: SVGMetadataElement;
+    SVGPathElement: SVGPathElement;
+    SVGRadialGradientElement: SVGRadialGradientElement;
+    SVGRectElement: SVGRectElement;
+    SVGSetElement: SVGSetElement;
+    SVGStopElement: SVGStopElement;
+    SVGSwitchElement: SVGSwitchElement;
+    SVGTSpanElement: SVGTSpanElement;
+    SVGTextContentElement: SVGTextContentElement;
+    SVGTextElement: SVGTextElement;
+    SVGTextPositioningElement: SVGTextPositioningElement;
+    SVGTitleElement: SVGTitleElement;
+    ScreenOrientation: ScreenOrientation;
+    ScriptProcessorNode: ScriptProcessorNode;
+    SecurityPolicyViolationEvent: SecurityPolicyViolationEvent;
+    ServiceWorkerContainer: ServiceWorkerContainer;
+    ServiceWorkerRegistration: ServiceWorkerRegistration;
+    SourceBuffer: SourceBuffer;
+    SourceBufferList: SourceBufferList;
+    SpeechSynthesis: SpeechSynthesis;
+    SpeechSynthesisErrorEvent: SpeechSynthesisErrorEvent;
+    SpeechSynthesisEvent: SpeechSynthesisEvent;
+    SpeechSynthesisUtterance: SpeechSynthesisUtterance;
+    StaticRange: StaticRange;
+    StereoPannerNode: StereoPannerNode;
+    StorageEvent: StorageEvent;
+    SubmitEvent: SubmitEvent;
+    TextDecoder: TextDecoder;
+    TextEncoder: TextEncoder;
+    TextTrack: TextTrack;
+    TextTrackCue: TextTrackCue;
+    TextTrackList: TextTrackList;
+    TouchEvent: TouchEvent;
+    TrackEvent: TrackEvent;
+    TransitionEvent: TransitionEvent;
+    UIEvent: UIEvent;
+    VTTCue: VTTCue;
+    VisualViewport: VisualViewport;
+    WaveShaperNode: WaveShaperNode;
+    WebGLContextEvent: WebGLContextEvent;
+    WebSocket: WebSocket;
+    WheelEvent: WheelEvent;
+    XMLDocument: XMLDocument;
+    XMLHttpRequest: XMLHttpRequest;
+    XMLHttpRequestEventTarget: XMLHttpRequestEventTarget;
+    XMLHttpRequestUpload: XMLHttpRequestUpload;
+    XPathEvaluator: XPathEvaluator;
+}
