+++ src/proxy/recursive/index.ts
@@ -1,9 +1,6 @@
-import { methods } from '../../methods';
-import { proxyRecursiveApply } from './apply.recursive.proxy.engine';
-import {
-  CommandMixType,
-  ParamsProxyEngineInterface,
-} from '../../interfaces/engine/proxy/params.proxy.engine.interface';
+import {predefinedMethods} from '../../methods';
+import {proxyRecursiveApply} from './apply.recursive.proxy.engine';
+import {CommandMixType, ParamsProxyEngineInterface,} from '../../interfaces/engine/proxy/params.proxy.engine.interface';
 
 type proxyRecursiveGetType = (targetGet: any, nameGet: string) => ReturnType<typeof proxyRecursive>;
 
@@ -25,7 +22,7 @@ export function proxyRecursive(target: object | string, name: string, params: Pa
 
   if (typeof target === 'string') {
     if (['not', 'or'].includes(target)) {
-      target = methods;
+      target = predefinedMethods;
     } else {
       target = () => {};
     }
@@ -38,8 +35,8 @@ export function proxyRecursive(target: object | string, name: string, params: Pa
 }
 
 function getMethod(name: string): CommandMixType | string {
-  if (Reflect.has(methods, name)) {
-    return (methods as any)[name];
+  if (Reflect.has(predefinedMethods, name)) {
+    return (predefinedMethods as any)[name];
   }
 
   return name;
